[APP]: Moving `User` model to `models` folder
